Databanks, Excel, Word files for CSRD Ordered
Please mind that the content of the CSRD is governed by the Cypher System Open License. The CC0 only applies to the work I did. If you ever use my work in your project (be it a sold product or something for the community), please mention me as Kaspar Schultess and notify me via email or github.